requirements.txt
# 🌞 HelioBio-Social - Dependencias Compatibles
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1

# Análisis de datos
pandas==2.1.4
numpy==1.24.4
scipy==1.10.1
scikit-learn==1.3.2

# Visualización
matplotlib==3.7.3
seaborn==0.13.0
plotly==5.17.0

# Utilidades
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2